Refactored member cleanup jobs to extract testable logic into lib/
- pulled the token and expired-comped cleanup logic out of the worker-thread
  entry files into lib/clean-tokens.js and lib/clean-expired-comped.js, each a
  plain async function taking db, so the logic is callable without a worker
- the worker files now just call the extracted function; behaviour unchanged
- prep for scheduling these via the in-memory JobQueue; the extracted lib is
  what the new handlers will reuse
- added a unit test for cleanTokens, which previously had only integration cover
